xc-recon-v2/backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# AI Integration
openai==1.3.7

# 3D File Processing
vtk==9.3.0
numpy==1.24.4
matplotlib==3.8.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Async Support
asyncio-mqtt==0.16.1
websockets==12.0

# File Upload
aiofiles==23.2.1